refactor: reverted packaging
